SAK-39982 Fix the stats for assignments to allow blank increments to be added without affecting the stats spread. Fix up the window title, the model getting confused with a null and a string arg Wire up the average GPA stats mMove the charts to a separate package Consolidate the properties Remove JFreechart entirely SAK-39982
