+++ gradebookng/tool/pom.xml
@@ -112,11 +112,6 @@
 			<scope>provided</scope> <!-- to satisfy EB and eclipse -->
 		</dependency>
 		<dependency>
-			<groupId>org.jfree</groupId>
-			<artifactId>jfreechart</artifactId>
-			<version>1.0.19</version>
-		</dependency>
-		<dependency>
 			<groupId>org.apache.commons</groupId>
 			<artifactId>commons-math3</artifactId>
 		</dependency>
