+++ gradebookng/tool/src/java/org/sakaiproject/gradebookng/GradebookNgApplication.properties
@@ -501,32 +501,21 @@ no-students.label = There are no students with Gradebook items
 
 gradebookpage.uncategorised = Uncategorized
 
-label.statistics.title = Grade Statistics for {0}
-label.statistics.average = Average (Mean) Score
-label.statistics.averagevalue = {0} / {1} ({2})
-label.statistics.graded = Total Graded Scores
-label.statistics.median = Median Score
-label.statistics.medianvalue = {0} / {1} ({2})
-label.statistics.lowest = Lowest Score
-label.statistics.lowestvalue = {0} / {1} ({2})
-label.statistics.highest = Highest Score
-label.statistics.highestvalue = {0} / {1} ({2})
-label.statistics.deviation = Standard Deviation
+label.statistics.title.assignment = Grade Statistics for {0}
+label.statistics.title.coursegrade = Course Grade Statistics
+label.statistics.average = Average (mean) grade
+label.statistics.graded = Total graded
+label.statistics.median = Median grade
+label.statistics.lowest = Lowest grade
+label.statistics.highest = Highest grade
+label.statistics.deviation = Standard deviation
+label.statistics.averagegpa = Course Average GPA
 label.statistics.chart.extracredit = EC
 label.statistics.chart.xaxis = Percentage Scored
 label.statistics.chart.yaxis = Number of Students
 label.statistics.chart.tooltip = {2} score(s) in {1} range
 label.statistics.chart.range={0}-{1}
 label.statistics.chart.title=Grade Distribution
-
-label.statistics.coursegrade.title = Course Grade Statistics
-label.statistics.coursegrade.averagegpa = Course Average GPA
-label.statistics.coursegrade.average = Average grade
-label.statistics.coursegrade.graded = Total graded students
-label.statistics.coursegrade.median = Median grade
-label.statistics.coursegrade.lowest = Lowest grade
-label.statistics.coursegrade.highest = Highest grade
-label.statistics.coursegrade.deviation = Standard deviation
 
 coursegrade.option.override = Course Grade Override
 coursegrade.option.overridelog = Course Grade Override Log
