+++ gradebookng/tool/src/java/org/sakaiproject/gradebookng/tool/panels/AssignmentStatisticsPanel.java
@@ -24,8 +24,8 @@ import org.apache.wicket.extensions.ajax.markup.html.modal.ModalWindow;
 import org.apache.wicket.model.IModel;
 import org.apache.wicket.model.Model;
 import org.apache.wicket.model.StringResourceModel;
+import org.sakaiproject.gradebookng.tool.chart.AssignmentGradeChart;
 import org.sakaiproject.gradebookng.tool.component.GbAjaxLink;
-import org.sakaiproject.gradebookng.tool.component.GbAssignmentGradeChart;
 import org.sakaiproject.gradebookng.tool.stats.AssignmentStatistics;
 import org.sakaiproject.service.gradebook.shared.Assignment;
 
@@ -49,9 +49,9 @@ public class AssignmentStatisticsPanel extends BasePanel {
 		final Assignment assignment = this.businessService.getAssignment(assignmentId.longValue());
 
 		AssignmentStatisticsPanel.this.window.setTitle(
-				new StringResourceModel("label.statistics.title", null, assignment.getName()).getString());
+				new StringResourceModel("label.statistics.title.assignment", getDefaultModel(), null, assignment.getName()).getString());
 
-		final GbAssignmentGradeChart chart = new GbAssignmentGradeChart("gradingSchemaChart", assignmentId);
+		final AssignmentGradeChart chart = new AssignmentGradeChart("gradingSchemaChart", assignmentId);
 		add(chart);
 
 		final AssignmentStatistics stats = new AssignmentStatistics("stats", getData(assignment));
